change to `Bean`
