+++ src/main/java/stirling/software/SPDF/controller/web/ConverterWebController.java
@@ -1,6 +1,5 @@
 package stirling.software.SPDF.controller.web;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,20 +9,15 @@ import org.springframework.web.servlet.ModelAndView;
 import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
-import stirling.software.SPDF.config.security.AppUpdateShowService;
-
 @Controller
 @Tag(name = "Convert", description = "Convert APIs")
 public class ConverterWebController {
-
-    @Autowired private AppUpdateShowService appUpdateShowService;
 
     @ConditionalOnExpression("#{bookAndHtmlFormatsInstalled}")
     @GetMapping("/book-to-pdf")
     @Hidden
     public String convertBookToPdfForm(Model model) {
         model.addAttribute("currentPage", "book-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/book-to-pdf";
     }
 
@@ -32,7 +26,6 @@ public class ConverterWebController {
     @Hidden
     public String convertPdfToBookForm(Model model) {
         model.addAttribute("currentPage", "pdf-to-book");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/pdf-to-book";
     }
 
@@ -40,7 +33,6 @@ public class ConverterWebController {
     @Hidden
     public String convertImgToPdfForm(Model model) {
         model.addAttribute("currentPage", "img-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/img-to-pdf";
     }
 
@@ -48,7 +40,6 @@ public class ConverterWebController {
     @Hidden
     public String convertHTMLToPdfForm(Model model) {
         model.addAttribute("currentPage", "html-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/html-to-pdf";
     }
 
@@ -56,7 +47,6 @@ public class ConverterWebController {
     @Hidden
     public String convertMarkdownToPdfForm(Model model) {
         model.addAttribute("currentPage", "markdown-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/markdown-to-pdf";
     }
 
@@ -64,7 +54,6 @@ public class ConverterWebController {
     @Hidden
     public String convertURLToPdfForm(Model model) {
         model.addAttribute("currentPage", "url-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/url-to-pdf";
     }
 
@@ -72,7 +61,6 @@ public class ConverterWebController {
     @Hidden
     public String pdfToimgForm(Model model) {
         model.addAttribute("currentPage", "pdf-to-img");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/pdf-to-img";
     }
 
@@ -80,7 +68,6 @@ public class ConverterWebController {
     @Hidden
     public String convertToPdfForm(Model model) {
         model.addAttribute("currentPage", "file-to-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/file-to-pdf";
     }
 
@@ -91,7 +78,6 @@ public class ConverterWebController {
     public ModelAndView pdfToHTML() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-html");
         modelAndView.addObject("currentPage", "pdf-to-html");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -100,7 +86,6 @@ public class ConverterWebController {
     public ModelAndView pdfToPresentation() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-presentation");
         modelAndView.addObject("currentPage", "pdf-to-presentation");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -109,7 +94,6 @@ public class ConverterWebController {
     public ModelAndView pdfToText() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-text");
         modelAndView.addObject("currentPage", "pdf-to-text");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -118,7 +102,6 @@ public class ConverterWebController {
     public ModelAndView pdfToWord() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-word");
         modelAndView.addObject("currentPage", "pdf-to-word");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -127,7 +110,6 @@ public class ConverterWebController {
     public ModelAndView pdfToXML() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-xml");
         modelAndView.addObject("currentPage", "pdf-to-xml");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -136,7 +118,6 @@ public class ConverterWebController {
     public ModelAndView pdfToCSV() {
         ModelAndView modelAndView = new ModelAndView("convert/pdf-to-csv");
         modelAndView.addObject("currentPage", "pdf-to-csv");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -144,7 +125,6 @@ public class ConverterWebController {
     @Hidden
     public String pdfToPdfAForm(Model model) {
         model.addAttribute("currentPage", "pdf-to-pdfa");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "convert/pdf-to-pdfa";
     }
 }
