+++ src/main/java/stirling/software/SPDF/controller/web/HomeWebController.java
@@ -20,20 +20,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.swagger.v3.oas.annotations.Hidden;
 
-import stirling.software.SPDF.config.security.AppUpdateShowService;
 import stirling.software.SPDF.model.ApplicationProperties;
 import stirling.software.SPDF.model.Dependency;
 
 @Controller
 public class HomeWebController {
 
-    @Autowired private AppUpdateShowService appUpdateShowService;
-
     @GetMapping("/about")
     @Hidden
     public String gameForm(Model model) {
         model.addAttribute("currentPage", "about");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "about";
     }
 
@@ -41,7 +37,6 @@ public class HomeWebController {
     @Hidden
     public String licensesForm(Model model) {
         model.addAttribute("currentPage", "licenses");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         Resource resource = new ClassPathResource("static/3rdPartyLicenses.json");
         try {
             InputStream is = resource.getInputStream();
@@ -59,7 +54,6 @@ public class HomeWebController {
     @GetMapping("/")
     public String home(Model model) {
         model.addAttribute("currentPage", "home");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "home";
     }
