+++ src/main/java/stirling/software/SPDF/controller/web/OtherWebController.java
@@ -6,7 +6,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,19 +14,14 @@ import org.springframework.web.servlet.ModelAndView;
 import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
-import stirling.software.SPDF.config.security.AppUpdateShowService;
-
 @Controller
 @Tag(name = "Misc", description = "Miscellaneous APIs")
 public class OtherWebController {
-
-    @Autowired private AppUpdateShowService appUpdateShowService;
 
     @GetMapping("/compress-pdf")
     @Hidden
     public String compressPdfForm(Model model) {
         model.addAttribute("currentPage", "compress-pdf");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/compress-pdf";
     }
 
@@ -36,7 +30,6 @@ public class OtherWebController {
     public ModelAndView extractImageScansForm() {
         ModelAndView modelAndView = new ModelAndView("misc/extract-image-scans");
         modelAndView.addObject("currentPage", "extract-image-scans");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -44,7 +37,6 @@ public class OtherWebController {
     @Hidden
     public String extractJavascriptForm(Model model) {
         model.addAttribute("currentPage", "show-javascript");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/show-javascript";
     }
 
@@ -52,7 +44,6 @@ public class OtherWebController {
     @Hidden
     public String stampForm(Model model) {
         model.addAttribute("currentPage", "stamp");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/stamp";
     }
 
@@ -60,7 +51,6 @@ public class OtherWebController {
     @Hidden
     public String addPageNumbersForm(Model model) {
         model.addAttribute("currentPage", "add-page-numbers");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/add-page-numbers";
     }
 
@@ -68,7 +58,6 @@ public class OtherWebController {
     @Hidden
     public String extractImagesForm(Model model) {
         model.addAttribute("currentPage", "extract-images");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/extract-images";
     }
 
@@ -76,7 +65,6 @@ public class OtherWebController {
     @Hidden
     public String flattenForm(Model model) {
         model.addAttribute("currentPage", "flatten");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/flatten";
     }
 
@@ -84,7 +72,6 @@ public class OtherWebController {
     @Hidden
     public String addWatermarkForm(Model model) {
         model.addAttribute("currentPage", "change-metadata");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/change-metadata";
     }
 
@@ -92,7 +79,6 @@ public class OtherWebController {
     @Hidden
     public String compareForm(Model model) {
         model.addAttribute("currentPage", "compare");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/compare";
     }
 
@@ -117,7 +103,6 @@ public class OtherWebController {
         Collections.sort(languages);
         modelAndView.addObject("languages", languages);
         modelAndView.addObject("currentPage", "ocr-pdf");
-        modelAndView.addObject("showUpdate", appUpdateShowService.showUpdate());
         return modelAndView;
     }
 
@@ -125,7 +110,6 @@ public class OtherWebController {
     @Hidden
     public String overlayImage(Model model) {
         model.addAttribute("currentPage", "add-image");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/add-image";
     }
 
@@ -133,7 +117,6 @@ public class OtherWebController {
     @Hidden
     public String contrast(Model model) {
         model.addAttribute("currentPage", "adjust-contrast");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/adjust-contrast";
     }
 
@@ -141,7 +124,6 @@ public class OtherWebController {
     @Hidden
     public String repairForm(Model model) {
         model.addAttribute("currentPage", "repair");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/repair";
     }
 
@@ -149,7 +131,6 @@ public class OtherWebController {
     @Hidden
     public String removeBlanksForm(Model model) {
         model.addAttribute("currentPage", "remove-blanks");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/remove-blanks";
     }
 
@@ -157,7 +138,6 @@ public class OtherWebController {
     @Hidden
     public String removeAnnotationsForm(Model model) {
         model.addAttribute("currentPage", "remove-annotations");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/remove-annotations";
     }
 
@@ -165,7 +145,6 @@ public class OtherWebController {
     @Hidden
     public String autoCropForm(Model model) {
         model.addAttribute("currentPage", "auto-crop");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/auto-crop";
     }
 
@@ -173,7 +152,6 @@ public class OtherWebController {
     @Hidden
     public String autoRenameForm(Model model) {
         model.addAttribute("currentPage", "auto-rename");
-        model.addAttribute("showUpdate", appUpdateShowService.showUpdate());
         return "misc/auto-rename";
     }
 }
